requirements.txt
setuptools==58
pyxdg==0.27
PyGObject==3.40.1
bluezero==0.6.0
dbus-python==1.2.18
